feat: WIP Add new Persons class for  persons parsing.
