+++ src/models/wikimedia/wikipedia/implementation_notes.md
@@ -1,0 +1,3 @@
+This directory contain all the logic specific to Wikipedia
+
+That includes article, references, multi-wiki support, etc.
