+++ src/models/wikimedia/wikipedia/reference/template/implementation_notes.md
@@ -1,0 +1,1 @@
+This directory holds all code that handles Wikipedia templates.
