+++ src/models/wikimedia/wikipedia/reference/template/template.py
@@ -345,69 +345,3 @@ class WikipediaTemplate(BaseModel):
         """Return a dict that we can output to patrons via the API"""
         return dict(parameters=self.parameters)
 
-    # todo enable parsing of persons
-    # def __parse_persons__(self) -> None:
-    #     """Parse all person related data into Person objects"""
-    #     # find all the attributes but exclude the properties as they lead to weird errors
-    #     properties = [
-    #         "has_hash",
-    #         "isodate",
-    #         "shortened_raw_template",
-    #         "template_url",
-    #         "wikibase_url",
-    #     ]
-    #     attributes = [
-    #         a
-    #         for a in dir(self)
-    #         if not a.startswith("_")
-    #         and a not in properties
-    #         and not callable(getattr(self, a))
-    #         and getattr(self, a) is not None
-    #     ]
-    #     self.authors_list = self.__parse_known_role_persons__(
-    #         attributes=attributes, role=EnglishWikipediaTemplatePersonRole.AUTHOR
-    #     )
-    #     self.editors_list = self.__parse_known_role_persons__(
-    #         attributes=attributes, role=EnglishWikipediaTemplatePersonRole.EDITOR
-    #     )
-    #     self.hosts_list = self.__parse_known_role_persons__(
-    #         attributes=attributes, role=EnglishWikipediaTemplatePersonRole.HOST
-    #     )
-    #     self.interviewers_list = self.__parse_known_role_persons__(
-    #         attributes=attributes, role=EnglishWikipediaTemplatePersonRole.INTERVIEWER
-    #     )
-    #     self.persons_without_role = self.__parse_roleless_persons__(
-    #         attributes=attributes
-    #     )
-    #     self.translators_list = self.__parse_known_role_persons__(
-    #         attributes=attributes, role=EnglishWikipediaTemplatePersonRole.TRANSLATOR
-    #     )
-
-    # @validate_arguments
-    # def __parse_roleless_persons__(self, attributes: List[str]) -> List[Person]:
-    #     persons = []
-    #     # first last
-    #     unnumbered_first_last = [
-    #         attribute
-    #         for attribute in attributes
-    #         if self.__find_number__(attribute) is None
-    #         and (attribute == "first" or attribute == "last")
-    #     ]
-    #     logger.debug(f"{len(unnumbered_first_last)} unnumbered first lasts found.")
-    #     if len(unnumbered_first_last) > 0:
-    #         person = Person(
-    #             role=EnglishWikipediaTemplatePersonRole.UNKNOWN,
-    #         )
-    #         for attribute in unnumbered_first_last:
-    #             # print(attribute, getattr(self, attribute))
-    #             if attribute == "first":
-    #                 person.given = self.first
-    #             if attribute == "last":
-    #                 person.surname = self.last
-    #         # console.print(person)
-    #         persons.append(person)
-    #         # exit()
-    #     # We use list comprehension to get the numbered persons to
-    #     # ease code maintentenance and easily support a larger range if necessary
-    #     persons.extend(self.__get_numbered_persons__(attributes=attributes))
-    #     return persons
