+++ tests/wikipedia/reference/test_person.py
@@ -4,7 +4,7 @@ from src.models.wikimedia.wikipedia.reference.enums import (
     EnglishWikipediaTemplatePersonRole,
     WikipediaTemplatePersonRole,
 )
-from src.models.wikimedia.wikipedia.reference.person import Person
+from src.models.wikimedia.wikipedia.reference.template.person import Person
 
 
 class TestPerson(TestCase):
